# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwiftLoad API - Video and file downloader
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MediaInfoRequest {
  /** URL to fetch info for */
  url: string;
}

export interface MediaFormat {
  id: string;
  label: string;
  /** mp4, mp3, or file */
  type: string;
  /** @nullable */
  quality?: string | null;
}

export interface MediaInfo {
  /** Detected platform (youtube, instagram, tiktok, twitter, facebook, direct) */
  platform: string;
  /** Media title */
  title: string;
  /**
     * Thumbnail URL
     * @nullable
     */
  thumbnail: string | null;
  /**
     * Duration in seconds
     * @nullable
     */
  duration?: number | null;
  formats: MediaFormat[];
}

export interface DownloadRequest {
  /** URL to download */
  url: string;
  /** mp4, mp3, or best */
  format: string;
}

export interface DownloadToken {
  /** Signed token to use with /api/download/stream?token=... */
  token: string;
  /** Suggested filename for the download */
  filename: string;
}

export interface ErrorResponse {
  error: string;
}