# lib/api-zod/src/generated/types/downloadRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwiftLoad API - Video and file downloader
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadRequest {
  /** URL to download */
  url: string;
  /** mp4, mp3, or best */
  format: string;
}