# lib/api-zod/src/generated/types/downloadToken.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwiftLoad API - Video and file downloader
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadToken {
  /** Signed token to use with /api/download/stream?token=... */
  token: string;
  /** Suggested filename for the download */
  filename: string;
}