# lib/api-zod/src/generated/types/mediaFormat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwiftLoad API - Video and file downloader
 * OpenAPI spec version: 0.1.0
 */

export interface MediaFormat {
  id: string;
  label: string;
  /** mp4, mp3, or file */
  type: string;
  /** @nullable */
  quality?: string | null;
}